refactor: handle fetching params operation in LegacyContent.ts
